Converts an interval object to a data frame representation.

Converts an interval object to a data frame representation.

# S3 method for class 'interval'
as.data.frame(x, ...)

Arguments

x

An interval object to be converted.

...

dditional arguments to be passed to as.data.frame().

Value

A data frame representation of the interval.

Examples

as.data.frame(inter_city)
#>               min.temp.jan max.temp.jan min.temp.fev max.temp.fev min.temp.mars
#> Berlin                  -3            2           -3            4             1
#> Bremen                  -1            3           -2            4             1
#> Franfurt                -1            3           -1            5             2
#> Hamburg                 -1            3           -2            4             1
#> Munich                  -4            2           -4            3             0
#> Stuttgart               -3            3           -3            4             1
#> Aberdeen                 0            6            1            6             2
#> Birmingham               1            6            1            6             2
#> Kirkwall                 2            6            2            6             3
#> Leeds                    3            7            2            7             4
#> Liverpool                2            7            2            7             3
#> Londres                  2            7            2            7             3
#> Manchester               2            7            2            7             3
#> Newcastle                1            5            1            6             3
#> Plymouth                 4            8            4            8             5
#> Linz                    -4            1           -3            3             1
#> Vienne                  -3            2           -2            4             2
#> Antwerp                  1            6            1            6             3
#> Brussels                 1            6            1            6             3
#> Plovdiv                 -2            4            1            7             3
#> Sofia                   -4            2           -3            4             1
#> Rijeka                   3           11            4           12             6
#> Zagreb                   5           11            5           12             7
#> Copenhague              -1            3           -2            2             0
#> Almeria                  8           17            9           17            11
#> Barcelone                4           13            5           14             7
#> Bilbao                   6           13            6           14             7
#> Madrid                   0           11            2           12             3
#> Malaga                   7           16            8           17             9
#> Seville                  6           16            7           17             8
#> Valence                  5           15            6           16             8
#> Ajaccio                  4           13            4           13             6
#> Bordeaux                 2            9            3           11             4
#> Brest                    4            9            4            9             5
#> Cannes                   4           13            4           12             7
#> Le Havre                 4            6            3            6             6
#> Le Mans                  2            7            2            8             4
#> Lille                    1            6            1            6             3
#> Lyon                     1            6            1            8             3
#> Marseille                3           11            3           12             6
#> Nice                     6           13            6           13             8
#> Paris                    1            6            1            7             3
#> Strasbourg              -1            4           -1            6             2
#> Alghero                  7           13            7           13             7
#> Cagliari                 6           14            6           14             8
#> Genes                    6           11            6           12             8
#> Milan                   -4            7           -3            8             1
#> Naples                   4           12            5           13             7
#> Palerme                 11           15           11           15            11
#> Pise                     2           11            3           12             5
#> Rome                     4           13            4           13             6
#> Venise                  -1            6            1            8             4
#> Amsterdam                1            5            0            6             3
#> Cracovie                -6            1           -4            2            -1
#> Poznan                  -3            1           -4            2             0
#> Varsovie                -4            1           -4            1            -1
#> Braga                    8           16            9           16            10
#> Porto                   14           18           14           18            14
#> Lisbonne                 7           14            8           15             9
#> Arad                    -4            2           -2            4             2
#> Bucarest                -5            2           -4            4             1
#> Cluj-Napoca             -6           -1           -5            2             0
#> Deva                    -4            1           -3            4             1
#> Moscou                 -12           -6          -11           -4            -6
#> St Petesbourg           -9           -4           -9           -4            -4
#> Istambul                 3            8            3            8             4
#> Izmit                    4           12            4           12             7
#> Kiev                    -9           -3           -8           -2            -3
#>               max.temp.mars min.temp.avr max.temp.avr min.temp.mai max.temp.mai
#> Berlin                    8            3           12            7           18
#> Bremen                    8            3           12            7           17
#> Franfurt                 10            4           13            8           19
#> Hamburg                   7            3           11            7           17
#> Munich                    9            2           12            7           17
#> Stuttgart                 9            3           12            7           17
#> Aberdeen                  8            3           10            6           13
#> Birmingham                9            3           12            6           16
#> Kirkwall                  7            3            8            6           11
#> Leeds                     9            5           11            8           16
#> Liverpool                 9            4           12            7           16
#> Londres                   9            4           11            7           15
#> Manchester                9            4           11            7           15
#> Newcastle                 8            3           10            6           13
#> Plymouth                 10            6           12            9           14
#> Linz                      9            4           13            8           19
#> Vienne                    9            5           14            9           19
#> Antwerp                   9            4           13            8           17
#> Brussels                  9            4           13            8           17
#> Plovdiv                  12            7           18           12           23
#> Sofia                     9            5           14            9           19
#> Rijeka                   14            9           17           13           22
#> Zagreb                   14           10           17           14           21
#> Copenhague                5            2            9            7           16
#> Almeria                  18           12           20           15           23
#> Barcelone                 7            8           17           12           20
#> Bilbao                   16            8           17           11           19
#> Madrid                   16            6           17            9           22
#> Malaga                   19           11           20           13           23
#> Seville                  21           10           22           13           26
#> Valence                  18            9           19           13           22
#> Ajaccio                  14            7           17           11           21
#> Bordeaux                 14            6           16           10           19
#> Brest                    11            6           12            8           15
#> Cannes                   14            9           16           13           20
#> Le Havre                  8            7           10           11           13
#> Le Mans                  12            6           14            9           18
#> Lille                     9            4           12            8           17
#> Lyon                     12            6           14           10           19
#> Marseille                14            8           17           12           21
#> Nice                     14            9           16           13           19
#> Paris                    11            6           14            9           18
#> Strasbourg               11            4           14            9           19
#> Alghero                  15            9           17           12           22
#> Cagliari                 16            9           18           13           22
#> Genes                    14           11           16           14           20
#> Milan                    13            4           17            9           21
#> Naples                   16            9           18           13           22
#> Palerme                  16           13           18           16           21
#> Pise                     14            7           17           11           22
#> Rome                     15            8           17           12           22
#> Venise                   12            8           16           13           21
#> Amsterdam                 9            4           12            8           16
#> Cracovie                  7            3           12            8           18
#> Poznan                    7            3           12            8           18
#> Varsovie                  7            3           12            8           18
#> Braga                    18           11           19           13           22
#> Porto                    18           14           18           16           20
#> Lisbonne                 17           11           18           12           21
#> Arad                     11            6           16           10           21
#> Bucarest                 10            6           17           11           22
#> Cluj-Napoca               9            4           14            9           19
#> Deva                     11            6           16           10           21
#> Moscou                    1            1            9            7           17
#> St Petesbourg             1            1            8            7           16
#> Istambul                 11            8           16           12           21
#> Izmit                    16           10           20           14           25
#> Kiev                      4            3           13            8           20
#>               min.temp.juin max.temp.juin min.temp.juil max.temp.juil
#> Berlin                   12            21            13            23
#> Bremen                   11            20            12            21
#> Franfurt                 12            22            14            24
#> Hamburg                  11            19            13            21
#> Munich                   10            20            12            22
#> Stuttgart                11            21            13            23
#> Aberdeen                  8            16            11            18
#> Birmingham                9            18            12            21
#> Kirkwall                  8            13            10            15
#> Leeds                    11            18            13            20
#> Liverpool                11            18            13            20
#> Londres                  11            18            12            19
#> Manchester               11            18            12            19
#> Newcastle                 9            16            11            18
#> Plymouth                 12            17            13            19
#> Linz                     12            21            13            23
#> Vienne                   13            22            15            25
#> Antwerp                  12            20            13            22
#> Brussels                 11            19            13            22
#> Plovdiv                  16            27            18            29
#> Sofia                    13            23            14            26
#> Rijeka                   17            26            19            29
#> Zagreb                   18            25            21            28
#> Copenhague               11            19            13            21
#> Almeria                  18            27            22            29
#> Barcelone                16            24            19            27
#> Bilbao                   14            22            16            25
#> Madrid                   13            28            16            32
#> Malaga                   17            27            20            29
#> Seville                  17            31            19            35
#> Valence                  17            26            20            29
#> Ajaccio                  14            24            17            27
#> Bordeaux                 13            23            15            26
#> Brest                    11            18            13            20
#> Cannes                   17            23            19            27
#> Le Havre                 13            16            16            18
#> Le Mans                  12            21            14            24
#> Lille                    11            19            13            22
#> Lyon                     14            23            16            27
#> Marseille                16            26            19            29
#> Nice                     17            23            19            26
#> Paris                    12            21            14            24
#> Strasbourg               12            22            14            24
#> Alghero                  16            26            18            29
#> Cagliari                 17            27            19            30
#> Genes                    18            23            21            27
#> Milan                    13            25            16            28
#> Naples                   17            26            19            29
#> Palerme                  20            25            23            28
#> Pise                     14            26            17            29
#> Rome                     16            25            19            28
#> Venise                   16            24            18            27
#> Amsterdam                11            19            13            21
#> Cracovie                 11            21            13            22
#> Poznan                   11            21            13            22
#> Varsovie                 11            21            13            23
#> Braga                    17            25            19            28
#> Porto                    18            22            19            24
#> Lisbonne                 15            25            17            28
#> Arad                     13            24            14            27
#> Bucarest                 14            26            16            28
#> Cluj-Napoca              12            22            13            23
#> Deva                     13            23            14            26
#> Moscou                   11            21            13            22
#> St Petesbourg            11            19            13            21
#> Istambul                 17            26            19            28
#> Izmit                    18            31            20            33
#> Kiev                     12            23            13            24
#>               min.temp.aout max.temp.aout min.temp.sep max.temp.sep
#> Berlin                   13            23           10           19
#> Bremen                   12            22            9           18
#> Franfurt                 13            24           11           20
#> Hamburg                  12            22            9           18
#> Munich                   12            23            9           19
#> Stuttgart                12            23            9           19
#> Aberdeen                 10            18            8           15
#> Birmingham               11            21            9           17
#> Kirkwall                 10            15            8           13
#> Leeds                    13            19           10           17
#> Liverpool                12            19           10           17
#> Londres                  12            19           10           17
#> Manchester               12            19           10           17
#> Newcastle                11            18            9           16
#> Plymouth                 13            19           12           17
#> Linz                     13            23           10           19
#> Vienne                   15            25           12           21
#> Antwerp                  13            22           11           19
#> Brussels                 13            22           11           19
#> Plovdiv                  17            28           14           26
#> Sofia                    14            26           11           22
#> Rijeka                   19            29           16           26
#> Zagreb                   21            28           18           24
#> Copenhague               12            21           10           16
#> Almeria                  22            30           21           28
#> Barcelone                19            28           17           26
#> Bilbao                   16            25           14           24
#> Madrid                   16            32           13           28
#> Malaga                   21            29           18           28
#> Seville                  19            35           18           32
#> Valence                  21            29           18           27
#> Ajaccio                  17            28           15           26
#> Bordeaux                 15            26           12           23
#> Brest                    13            20           11           18
#> Cannes                   19            27           16           24
#> Le Havre                 16            19           14           17
#> Le Mans                  14            24           11           21
#> Lille                    13            23           11           19
#> Lyon                     16            26           12           22
#> Marseille                18            28           16           25
#> Nice                     20            27           17           24
#> Paris                    14            24           11           21
#> Strasbourg               13            24           11           21
#> Alghero                  18            29           17           26
#> Cagliari                 19            30           17           27
#> Genes                    21            27           18           24
#> Milan                    16            28           12           24
#> Naples                   19            29           16           27
#> Palerme                  24            29           22           27
#> Pise                     17            29           15           26
#> Rome                     19            28           17           26
#> Venise                   18            27           15           23
#> Amsterdam                13            21           11           18
#> Cracovie                 12            22            9           18
#> Poznan                   12            23            9           18
#> Varsovie                 12            23            8           18
#> Braga                    19            28           18           27
#> Porto                    21            25           21           24
#> Lisbonne                 18            28           17           27
#> Arad                     14            27           11           23
#> Bucarest                 15            28           11           24
#> Cluj-Napoca              13            24            9           20
#> Deva                     14            26           11           22
#> Moscou                   11            20            6           14
#> St Petesbourg            12            19            7           13
#> Istambul                 19            28           16           24
#> Izmit                    19            32           16           29
#> Kiev                     12            23            8           19
#>               min.temp.oct max.temp.oct min.temp.nov max.temp.nov min.temp.dec
#> Berlin                   6           13            2            7           -1
#> Bremen                   6           13            3            8            1
#> Franfurt                 6           14            2            7            0
#> Hamburg                  6           13            3            7            0
#> Munich                   4           13            0            7           -3
#> Stuttgart                5           14            1            8           -2
#> Aberdeen                 6           12            3            8            2
#> Birmingham               6           13            3            9            2
#> Kirkwall                 7           11            4            8            3
#> Leeds                    8           13            5            9            4
#> Liverpool                7           13            4            9            3
#> Londres                  7           13            4            9            3
#> Manchester               7           13            4            9            3
#> Newcastle                7           12            4            8            2
#> Plymouth                 9           14            7           11            6
#> Linz                     5           13            1            6           -2
#> Vienne                   6           14            2            7           -1
#> Antwerp                  7           14            4            9            2
#> Brussels                 8           14            4            9            2
#> Plovdiv                  8           18            3           11            0
#> Sofia                    7           16            1            8           -2
#> Rijeka                  12           21            8           16            5
#> Zagreb                  13           21            9           15            6
#> Copenhague               7           12            3            7            0
#> Almeria                 16           24           12           20            9
#> Barcelone               12           21            8           17            6
#> Bilbao                  12           21            9           17            7
#> Madrid                   8           20            4           14            2
#> Malaga                  14           23           11           19            9
#> Seville                 13           26            9           21            7
#> Valence                 13           23            9           18            7
#> Ajaccio                 12           22            8           17            5
#> Bordeaux                 9           18            5           13            3
#> Brest                    9           15            7           12            6
#> Cannes                  13           20            8           16            6
#> Le Havre                11           14            7            9            5
#> Le Mans                  8           16            4           11            3
#> Lille                    7           14            4            9            2
#> Lyon                     8           16            4           10            2
#> Marseille               11           20            7           14            3
#> Nice                    13           20            9           16            6
#> Paris                    8           15            4            9            2
#> Strasbourg               7           14            2            8            0
#> Alghero                 14           22           10           17            7
#> Cagliari                14           23            9           18            7
#> Genes                   14           19            9           14            7
#> Milan                    7           18            1           11           -3
#> Naples                  13           22            8           17            6
#> Palerme                 19           23           14           19           12
#> Pise                    11           21            6           15            3
#> Rome                    13           22            8           17            6
#> Venise                  10           28            4           11            1
#> Amsterdam                8           14            4            9            2
#> Cracovie                 4           13            0            6           -3
#> Poznan                   5           13            1            6           -2
#> Varsovie                 4           12            1            6           -3
#> Braga                   14           22           12           19            9
#> Porto                   18           23           17           21           16
#> Lisbonne                14           21           11           17            9
#> Arad                     6           17            1            8           -2
#> Bucarest                 6           17            1            8           -3
#> Cluj-Napoca              4           14           -1            6           -4
#> Deva                     6           16            1            7           -2
#> Moscou                   1            7           -4            0           -9
#> St Petesbourg            3            7           -2            2           -7
#> Istambul                13           19            8           14            6
#> Izmit                   13           23            8           17            6
#> Kiev                     3           12           -2            3           -6
#>               max.temp.dec min.humid max.humid
#> Berlin                   3        27        53
#> Bremen                   5        30        54
#> Franfurt                 4        30        54
#> Hamburg                  4        30        54
#> Munich                   3        26        53
#> Stuttgart                4        28        53
#> Aberdeen                 7        33        51
#> Birmingham               7        34        52
#> Kirkwall                 7        35        51
#> Leeds                    8        34        51
#> Liverpool                8        35        53
#> Londres                  7        34        52
#> Manchester               7        34        52
#> Newcastle                6        33        51
#> Plymouth                10        38        55
#> Linz                     2        26        54
#> Vienne                   4        26        54
#> Antwerp                  7        32        56
#> Brussels                 7        32        55
#> Plovdiv                  6        28        59
#> Sofia                    3        23        53
#> Rijeka                  12        32        57
#> Zagreb                  12        33        58
#> Copenhague               4        28        52
#> Almeria                 17        44        66
#> Barcelone               14        39        66
#> Bilbao                  14        39        60
#> Madrid                  11        34        52
#> Malaga                  17        44        64
#> Seville                 17        43        59
#> Valence                 16        40        66
#> Ajaccio                 14        41        64
#> Bordeaux                10        38        59
#> Brest                   10        40        56
#> Cannes                  13        25        85
#> Le Havre                 7        38        58
#> Le Mans                  8        36        57
#> Lille                    7        33        55
#> Lyon                     7        33        57
#> Marseille               12        36        59
#> Nice                    13        35        64
#> Paris                    7        33        55
#> Strasbourg               5        31        56
#> Alghero                 14        44        64
#> Cagliari                15        44        66
#> Genes                   12        34        64
#> Milan                    7        28        62
#> Naples                  13        40        70
#> Palerme                 16        46        70
#> Pise                    12        36        63
#> Rome                    14        40        67
#> Venise                   7        32        63
#> Amsterdam                7        33        56
#> Cracovie                 2        24        55
#> Poznan                   3        27        53
#> Varsovie                 2        25        55
#> Braga                   17        46        60
#> Porto                   19        52        63
#> Lisbonne                14        44        58
#> Arad                     3        28        59
#> Bucarest                 3        26        61
#> Cluj-Napoca              1        22        56
#> Deva                     2        25        58
#> Moscou                  -4        12        54
#> St Petesbourg           -2        16        54
#> Istambul                10        34        61
#> Izmit                   13        38        61
#> Kiev                     0        17        55