Results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars, aggregate on sulfur dioxide to interval simple label data.

Results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars, aggregate on sulfur dioxide to interval simple label data.

inter_wine

Format

A interval structure with 132 rows and 10 variables divided in min and max with 7 class:

fixed.acidity

volatile.acidity

citric.acid

residual.sugar

chlorides

total.sulfur.dioxide

density

pH

sulphates

alcohol

Class :

Class3

Class4

Class5

Class6

Class7

Class8

Class9

Source

https://archive.ics.uci.edu/dataset/186/wine+quality