OKM clustering.

Overlapping Kmeans algorithm.

okm(X, centers, iter.max = 10, nstart = 1, trace = FALSE, method = "euclid")

Arguments

X

A data.

centers

A number or matrix, number of cluster for clustering or pre init centers.

iter.max

the maximum number of iterations allowed.

nstart

A number, number of execution to find the best result.

trace

A boolean, tracing information on the progress of the algorithm is produced.

method

A string ('euclid': Euclidian distance, 'manhattan': Manhattan distance).

Examples

okm(iris[, -5], 3)
#> OKM clustering with 0 clusters of sizes 
#> 
#> Cluster means:
#>      Sepal.Length Sepal.Width Petal.Length Petal.Width
#> [1,]     7.038559    3.137736     5.958983   2.1611487
#> [2,]     4.993442    3.407433     1.473706   0.2546202
#> [3,]     5.652211    2.632185     4.134094   1.3109983
#> 
#> Clustering matrix:
#>        [,1] [,2] [,3]
#>   [1,]    0    1    0
#>   [2,]    0    1    0
#>   [3,]    0    1    0
#>   [4,]    0    1    0
#>   [5,]    0    1    0
#>   [6,]    0    1    0
#>   [7,]    0    1    0
#>   [8,]    0    1    0
#>   [9,]    0    1    0
#>  [10,]    0    1    0
#>  [11,]    0    1    0
#>  [12,]    0    1    0
#>  [13,]    0    1    0
#>  [14,]    0    1    0
#>  [15,]    0    1    0
#>  [16,]    0    1    0
#>  [17,]    0    1    0
#>  [18,]    0    1    0
#>  [19,]    0    1    0
#>  [20,]    0    1    0
#>  [21,]    0    1    0
#>  [22,]    0    1    0
#>  [23,]    0    1    0
#>  [24,]    0    1    0
#>  [25,]    0    1    0
#>  [26,]    0    1    0
#>  [27,]    0    1    0
#>  [28,]    0    1    0
#>  [29,]    0    1    0
#>  [30,]    0    1    0
#>  [31,]    0    1    0
#>  [32,]    0    1    0
#>  [33,]    0    1    0
#>  [34,]    0    1    0
#>  [35,]    0    1    0
#>  [36,]    0    1    0
#>  [37,]    0    1    0
#>  [38,]    0    1    0
#>  [39,]    0    1    0
#>  [40,]    0    1    0
#>  [41,]    0    1    0
#>  [42,]    0    1    0
#>  [43,]    0    1    0
#>  [44,]    0    1    0
#>  [45,]    0    1    0
#>  [46,]    0    1    0
#>  [47,]    0    1    0
#>  [48,]    0    1    0
#>  [49,]    0    1    0
#>  [50,]    0    1    0
#>  [51,]    1    0    1
#>  [52,]    1    0    1
#>  [53,]    1    0    1
#>  [54,]    0    0    1
#>  [55,]    1    0    1
#>  [56,]    0    0    1
#>  [57,]    1    0    1
#>  [58,]    0    1    1
#>  [59,]    1    0    1
#>  [60,]    0    0    1
#>  [61,]    0    0    1
#>  [62,]    0    0    1
#>  [63,]    0    0    1
#>  [64,]    1    0    1
#>  [65,]    1    1    1
#>  [66,]    1    0    1
#>  [67,]    0    0    1
#>  [68,]    0    0    1
#>  [69,]    0    0    1
#>  [70,]    0    0    1
#>  [71,]    1    0    1
#>  [72,]    0    0    1
#>  [73,]    1    0    1
#>  [74,]    1    0    1
#>  [75,]    0    0    1
#>  [76,]    1    0    1
#>  [77,]    1    0    1
#>  [78,]    1    0    1
#>  [79,]    0    0    1
#>  [80,]    1    1    1
#>  [81,]    0    0    1
#>  [82,]    0    0    1
#>  [83,]    0    0    1
#>  [84,]    1    0    1
#>  [85,]    0    0    1
#>  [86,]    1    1    1
#>  [87,]    1    0    1
#>  [88,]    0    0    1
#>  [89,]    0    0    1
#>  [90,]    0    0    1
#>  [91,]    0    0    1
#>  [92,]    1    0    1
#>  [93,]    0    0    1
#>  [94,]    0    1    1
#>  [95,]    0    0    1
#>  [96,]    0    0    1
#>  [97,]    0    0    1
#>  [98,]    0    0    1
#>  [99,]    0    1    1
#> [100,]    0    0    1
#> [101,]    1    0    0
#> [102,]    1    0    1
#> [103,]    1    0    0
#> [104,]    1    0    1
#> [105,]    1    0    0
#> [106,]    1    0    0
#> [107,]    0    0    1
#> [108,]    1    0    0
#> [109,]    1    0    0
#> [110,]    1    0    0
#> [111,]    1    0    1
#> [112,]    1    0    1
#> [113,]    1    0    0
#> [114,]    1    0    1
#> [115,]    1    0    1
#> [116,]    1    0    1
#> [117,]    1    0    1
#> [118,]    1    0    0
#> [119,]    1    0    0
#> [120,]    1    0    1
#> [121,]    1    0    0
#> [122,]    1    0    1
#> [123,]    1    0    0
#> [124,]    1    0    1
#> [125,]    1    0    0
#> [126,]    1    0    0
#> [127,]    1    0    1
#> [128,]    1    0    1
#> [129,]    1    0    1
#> [130,]    1    0    0
#> [131,]    1    0    0
#> [132,]    1    0    0
#> [133,]    1    0    1
#> [134,]    1    0    1
#> [135,]    1    0    1
#> [136,]    1    0    0
#> [137,]    1    0    0
#> [138,]    1    0    1
#> [139,]    1    0    1
#> [140,]    1    0    0
#> [141,]    1    0    0
#> [142,]    1    0    1
#> [143,]    1    0    1
#> [144,]    1    0    0
#> [145,]    1    0    0
#> [146,]    1    0    1
#> [147,]    1    0    1
#> [148,]    1    0    1
#> [149,]    1    0    1
#> [150,]    1    0    1
#> 
#> Within cluster sum of squares by elements:
#> NULL
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "tot.withinss" "overlaps"